Project-Study-Mate-2.0/requirements.txt
streamlit
pypdf
sentence-transformers
numpy
scikit-learn
python-dotenv
google-generative-ai